+++ docs/docs/resources/changelog.md
@@ -15,6 +15,7 @@ During the alpha period, things might break! We take breaking changes very serio
 ## Next
 
 - Added an orientation example to the native radio docs
+- Added the `tag-counter` part to `<wa-select>` to allow targeting the tag that shows when more than the max number of visible items has been selected
 - Fixed a number of broken event listeners throughout the docs
 
 ## 3.0.0-alpha.10